fix hydration mismatch in SideMenu links
