+++ src/modules/auth/application/useAuthUI.tsx
@@ -1,0 +1,3 @@
+export function useAuthUI() {
+  return {};
+}
