+++ src/modules/auth/domain/auth-model.ts
@@ -1,0 +1,4 @@
+export interface ICredentials {
+  email: string;
+  password: string;
+}
