+++ src/modules/dashboard/application/useDashboardUI.tsx
@@ -1,0 +1,3 @@
+export function useDashboardUI() {
+  return {};
+}
