+++ src/modules/dashboard/index.tsx
@@ -1,0 +1,3 @@
+export { DashboardView } from './ui/DashboardView';
+export type { IDashboardStat, IDashboardFeature } from './domain/dashboard-model';
+
