+++ src/pages/auth/index.tsx
@@ -1,11 +1,10 @@
-import AuthLayout from "./_layout";
-import AuthModule from "@/modules/auth";
+import AuthLayout from './_layout';
+import { AuthView } from '@/modules/auth';
 
 export default function Login() {
-
   return (
     <AuthLayout>
-      <AuthModule />
+      <AuthView />
     </AuthLayout>
   );
 }
